correcting warning

+++ program1.cpp
@@ -22,7 +22,7 @@ int main(int argc, char** argv){
 	getline(cin, arg);
 	if(arg.compare("AUTO")==0){
 		cout << "Auto-generating..." << endl;
-		arg = string('s', BUFSIZE+1);
+		arg = string('s', ((size_t) BUFSIZE)+1);
 	}
 	cout << "User input: " << arg << endl;
